fix(rag): 修复检索服务中的数据类型转换和文档ID匹配问题
- 将MemoryRetrieverService中score字段强制转换为float类型以确保数据一致性
- 修复MemoryVectorStoreService中删除逻辑的条件判断，将docId比较改为chunkId比较
- 确保向量存储服务能够正确删除对应的文档块而不是文档ID匹配项

+++ bootstrap/src/main/java/com/nageoffer/ai/ragent/rag/core/retrieve/MemoryRetrieverService.java
@@ -99,7 +99,7 @@ public class MemoryRetrieverService implements RetrieverService {
                 .map(scoredChunk -> new RetrievedChunk(
                         scoredChunk.getChunk().getChunkId(),
                         scoredChunk.getChunk().getContent(),
-                        scoredChunk.getScore()
+                        (float) scoredChunk.getScore()
                 ))
                 .collect(Collectors.toList());
     }
